Revert "Refactor the pose cache out of BodyScale into its own static class"
This reverts commit 098be9f2e7a7724eb2d63515d8edeed5fa5b0be6.
